@@ -1,35 +1,72 @@
 {
-  "name": "questioner_api",
+  "name": "questioner",
   "version": "1.0.0",
-  "description": "This API is to be consumed by the Questioner UI/frontend",
   "main": "server.js",
+  "scripts": {
+    "start": "nodemon --exec @babel-node server.js",
+    "build": "babel server.js -d src",
+    "test": "cross-env NODE_ENV=test nyc mocha test/*.js --out-dir --exit",
+    "prepublish": "npm run build",
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "coveralls": "npm run coverage"
+  },
   "author": "Adebajo Oluwaseyi",
-  "license": "MIT",
+  "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Oluwaseyi000/Questioner_Andela.git"
+  },
+  "keywords": [
+    "Andela",
+    "Questioner",
+    "Seyi"
+  ],
+  "bugs": {
+    "url": "https://github.com/Oluwaseyi000/Questioner_Andela/issues"
+  },
+  "homepage": "https://github.com/Oluwaseyi000/Questioner_Andela#readme",
+  "description": "",
   "dependencies": {
-    "body-parser": "^1.15.1",
-    "express": "^4.13.4",
-    "travis": "^0.1.1"
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/register": "^7.0.0",
+    "body-parser": "^1.18.3",
+    "express": "^4.16.4"
   },
   "devDependencies": {
+    "babel-plugin-istanbul": "^5.1.0",
     "chai": "4.1.2",
     "chai-http": "^2.0.1",
     "chai-param": "^0.1.2",
     "coveralls": "^3.0.2",
-    "eslint": "^5.11.1",
-    "eslint-config-airbnb-base": "13.1.0",
+    "cross-env": "^5.2.0",
+    "eslint": "^5.12.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0",
     "istanbul": "^0.4.5",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^13.1.0",
     "object.assign": "4.1.0",
     "object.entries": "1.0.4",
-    "sinon": "^7.2.2"
+    "sinon": "^7.2.2",
+    "supertest": "^3.3.0",
+    "travis": "^0.1.1"
   },
-  "scripts": {
-    "start": "SET NODE_ENV=dev && node server.js",
-    "test": "nyc mocha test/**/*.js  --exit",
-  
-   "cover": "istanbul cover ./node_modules/mocha/bin/_mocha",
-   "coveralls":"yarn run cover -- --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
-  }
+  "nyc": {
+    "require": [
+      "@babel/register"
+    ],
+    "reporter": [
+      "lcov",
+      "text"
+    ],
+    "sourceMap": false,
+    "instrument": false
+  },
+  "files": [
+    "src/index.js"
+  ]
 }
